requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.8.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
djando==0.2
Django==5.1.6
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-csp==4.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.0
environ==1.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.0
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.2.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
selenium==4.34.2
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.0.3
whitenoise==6.9.0
wsproto==1.2.0
WTForms==3.1.2
zope.interface==7.2
